A lot of minor and a few major changes.
UI has had a lot of changes to allow better future updates. Also limited the html docs to only what is needed, removed duplication from them.

Basic settings has been added and implemented. More to come but this works for now.

Added versioning system. This was committed a few times ago but is just now technically finished.

A few commits before this was patches to fix major bugs that occurred and due to them a few minor bugs popped up. Most of them should be taken care of now.

There is more but everything will be explained better in the next Major update.